+++ src/main/java/no/entur/antu/commondata/CommonDataRepository.java
@@ -2,31 +2,9 @@ package no.entur.antu.commondata;
 
 import no.entur.antu.stop.model.QuayId;
 
-import java.util.Map;
+public interface CommonDataRepository {
 
-public class CommonDataRepository {
-
-    private final CommonDataResource commonDataResource;
-    private final Map<String, QuayId> quayIdsPerScheduledStopPointsCache;
-
-    public CommonDataRepository(CommonDataResource commonDataResource,
-                                Map<String, QuayId> quayIdsPerScheduledStopPointsCache) {
-        this.commonDataResource = commonDataResource;
-        this.quayIdsPerScheduledStopPointsCache = quayIdsPerScheduledStopPointsCache;
-    }
-
-    public QuayId findQuayId(String scheduledStopPoint) {
-        return quayIdsPerScheduledStopPointsCache.get(scheduledStopPoint);
-    }
-
-    public void loadCommonDataCache(byte[] fileContent) {
-        commonDataResource.loadCommonData(fileContent);
-        quayIdsPerScheduledStopPointsCache.putAll(
-                commonDataResource.getQuayIdsPerScheduledStopPoints()
-        );
-    }
-
-    public void cleanUp() {
-        quayIdsPerScheduledStopPointsCache.clear();
-    }
+    QuayId findQuayId(String scheduledStopPoint);
+    void loadCommonDataCache(byte[] fileContent);
+    void cleanUp();
 }
