+++ src/main/java/no/entur/antu/config/CommonDataConfig.java
@@ -1,6 +1,7 @@
 package no.entur.antu.config;
 
 import no.entur.antu.commondata.CommonDataRepository;
+import no.entur.antu.commondata.CommonDataRepositoryImpl;
 import no.entur.antu.commondata.CommonDataResource;
 import no.entur.antu.stop.model.QuayId;
 import org.redisson.api.RLocalCachedMap;
@@ -25,6 +26,6 @@ public class CommonDataConfig {
     CommonDataRepository commonDataRepository(CommonDataResource commonDataResource,
                                               @Qualifier(QUAY_IDS_PER_SCHEDULED_STOP_POINTS_CACHE)
                                               RLocalCachedMap<String, QuayId> quayIdsPerScheduledStopPointsCache) {
-        return new CommonDataRepository(commonDataResource, quayIdsPerScheduledStopPointsCache);
+        return new CommonDataRepositoryImpl(commonDataResource, quayIdsPerScheduledStopPointsCache);
     }
 }
