+++ src/main/java/no/entur/antu/routes/stop/StopRouteBuilder.java
@@ -39,11 +39,11 @@ public class StopRouteBuilder extends BaseRouteBuilder {
     public void configure() throws Exception {
         super.configure();
 
-/*        from("master:lockOnAntuRefreshStopCacheAtStartup:quartz://antu/refreshStopPlaceCacheAtStartup?" + "?trigger.repeatCount=0")
+        from("quartz://antu/refreshStopPlaceCacheAtStartup?" + "?trigger.repeatCount=0")
                 .log(LoggingLevel.INFO, correlation() + "Refreshing stop place cache at startup")
                 .to("direct:refresh-stop-cache")
                 .routeId("refresh-stop-cache-at-startup");
-*/
+
         from("master:lockOnAntuRefreshStopCachePeriodically:quartz://antu/refreshStopPlaceCachePeriodically?" + quartzTrigger)
                 .log(LoggingLevel.INFO, correlation() + "Refreshing stop place cache periodically")
                 .to("direct:refresh-stop-cache")
